Create clustering templates for active standby
Issues:
Fixes #76

Problem:
Two templates are needed, one to create clusterable VEs and one to
actually cluster two devices.

Analysis:
Included here are the two templates for clustering and the functional
tests. It is important to note that this will not merge until the
clustering feature in f5-sdk is merged from development into 0.1.

Tests:
Two additional functional tests exist. One for the creation of two
clusterable VEs and one for implementing the actual cluster.